refactor: replace local E2E harness with cardano-node-clients:devnet
Delete local Devnet.hs (280 lines) and Setup.hs (195 lines), replacing
them with the shared devnet library from cardano-node-clients. Updates
imports in all 7 spec files and removes 5 unused deps.

Closes #72
