+++ cabal.project
@@ -57,7 +57,8 @@ source-repository-package
 source-repository-package
   type: git
   location: https://github.com/paolino/cardano-node-clients
-  tag: 7677065cbcadaa784e60e6974b552a9cd7ef0dc3
+  tag: 6586effe864e0a6f28d0df36210958b62c4e3f39
+  --sha256: 09bsmrp8pgnqzxx2hxw9b3pdv0p0zzqiji43wxw7ha8b6d0igivv
 
 -- Enable iohk flag for contra-tracer-contrib to use CHaP's simple API
 constraints: contra-tracer-contrib +iohk
