+++ cardano-mpfs-offchain/e2e-test/Cardano/MPFS/E2E/CageFlowSpec.hs
@@ -71,17 +71,6 @@ import Cardano.MPFS.Core.Types
     , TokenId (..)
     , TokenState (..)
     )
-import Cardano.MPFS.E2E.Devnet
-    ( withCardanoNode
-    )
-import Cardano.MPFS.E2E.Setup
-    ( addKeyWitness
-    , devnetMagic
-    , genesisAddr
-    , genesisDir
-    , genesisSignKey
-    , keyHashFromSignKey
-    )
 import Cardano.MPFS.Provider (Provider (..))
 import Cardano.MPFS.State
     ( Requests (..)
@@ -99,6 +88,17 @@ import Cardano.MPFS.TxBuilder.Config
 import Cardano.MPFS.TxBuilder.Real.Internal
     ( cagePolicyIdFromCfg
     , computeScriptHash
+    )
+import Cardano.Node.Client.E2E.Devnet
+    ( withCardanoNode
+    )
+import Cardano.Node.Client.E2E.Setup
+    ( addKeyWitness
+    , devnetMagic
+    , genesisAddr
+    , genesisDir
+    , genesisSignKey
+    , keyHashFromSignKey
     )
 
 -- | CageFlow E2E test spec.
